requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
colored==1.3.93
cryptography==2.4.2
future==0.17.1
idna==2.8
pycparser==2.19
python-telegram-bot==11.1.0
six==1.12.0
